+++ features/ui/Button/index.ts
@@ -1,1 +1,1 @@
-export { Button } from "./button";
+export { Button } from "./abutton";
